Live for Demo/requirements.txt
faiss-cpu==1.10.0
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.39
langchain-openai==0.3.7
langchain-deepseek==0.1.2
langchain-text-splitters==0.3.6
openai==1.65.2
python-dotenv==1.0.1
streamlit==1.42.0
psycopg==3.2.6
psycopg-binary==3.2.6